refactor: Update Bagel configuration and precision handling
- Changed `run_name` in `bagel_example.yaml` from `blip3o` to `bagel` for clarity.
- Increased `max_latent_size` from 32 to 64 in `bagel_collator.py`, `bagel.py`, and related classes to enhance model capacity.
- Updated `PositionEmbedding` to accept a `dtype` parameter for improved precision handling across various model components.
- Added dtype checks and conversions in multiple locations to ensure consistent tensor precision throughout the Bagel model.

+++ src/lmms_engine/models/bagel/modeling_utils.py
@@ -134,12 +134,12 @@ class MLPconnector(nn.Module):
 
 
 class PositionEmbedding(nn.Module):
-    def __init__(self, max_num_patch_per_side, hidden_size):
+    def __init__(self, max_num_patch_per_side, hidden_size, dtype=torch.float32):
         super().__init__()
         self.max_num_patch_per_side = max_num_patch_per_side
         self.hidden_size = hidden_size
         self.pos_embed = nn.Parameter(
-            torch.zeros(max_num_patch_per_side**2, hidden_size), requires_grad=False
+            torch.zeros(max_num_patch_per_side**2, hidden_size, dtype=dtype), requires_grad=False
         )
         self._init_weights()
 
@@ -148,7 +148,7 @@ class PositionEmbedding(nn.Module):
         pos_embed = get_2d_sincos_pos_embed(
             self.hidden_size, self.max_num_patch_per_side
         )
-        self.pos_embed.data.copy_(torch.from_numpy(pos_embed).float())
+        self.pos_embed.data.copy_(torch.from_numpy(pos_embed).float().to(self.pos_embed.dtype))
 
     def forward(self, position_ids):
         return self.pos_embed[position_ids]
